+++ R/tm_t_events_summary.R
@@ -70,6 +70,19 @@ template_events_summary <- function(anl_name,
     )
   )
 
+  # Since this is a compound table with one layout based on `parentname`
+  # and one layout on `dataname`, columns will be filtered to match levels
+  # present in `parentname` only so `drop_arm_levels` = FALSE.
+  data_list <- add_expr(
+    data_list,
+    prepare_arm_levels(
+      dataname = "anl",
+      parentname = parentname,
+      arm_var = arm_var,
+      drop_arm_levels = FALSE
+    )
+  )
+
   data_list <- add_expr(
     data_list,
     quote(study_id <- unique(anl[["STUDYID"]]))
@@ -539,9 +552,9 @@ template_events_summary <- function(anl_name,
 #'       TMPFL_SER = "Serious AE",
 #'       TMPFL_REL = "Related AE",
 #'       TMPFL_GR5 = "Grade 5 AE",
-#'       TMP_SMQ01 = aesi_label(ADAE$SMQ01NAM, ADAE$SMQ01SC),
+#'       TMP_SMQ01 = aesi_label(dat[["SMQ01NAM"]], dat[["SMQ01SC"]]),
 #'       TMP_SMQ02 = aesi_label("Y.9.9.9.9/Z.9.9.9.9 AESI"),
-#'       TMP_CQ01 = aesi_label(ADAE$CQ01NAM)
+#'       TMP_CQ01 = aesi_label(dat[["CQ01NAM"]])
 #'     )
 #' }
 #'
@@ -580,9 +593,9 @@ template_events_summary <- function(anl_name,
 #'                 TMPFL_SER = "Serious AE",
 #'                 TMPFL_REL = "Related AE",
 #'                 TMPFL_GR5 = "Grade 5 AE",
-#'                 TMP_SMQ01 = aesi_label(ADAE$SMQ01NAM, ADAE$SMQ01SC),
+#'                 TMP_SMQ01 = aesi_label(dat[["SMQ01NAM"]], dat[["SMQ01SC"]]),
 #'                 TMP_SMQ02 = aesi_label("Y.9.9.9.9/Z.9.9.9.9 AESI"),
-#'                 TMP_CQ01 = aesi_label(ADAE$CQ01NAM)
+#'                 TMP_CQ01 = aesi_label(dat[["CQ01NAM"]])
 #'               )
 #'           }
 #'           # Generating user-defined event flags.