+++ NEWS.md
@@ -5,6 +5,7 @@
 * New module `tm_g_ipp` was added for individual patient plots.
 
 ### Enhancements
+* The argument `drop_arm_levels` was added to all safety modules. This allows removal of columns based on factor levels not found in filtered data.
 * `tm_g_km` updated to allow plot of failure probability on y-axis and tick interval selection on x-axis.
 * The argument `time_unit_var` was added to `template_g_km` to add the time unit to the x-axis label.